[add-新增功能] 功能描述:添加jvm内存查询 影响范围: StoryID :

+++ dfas-monitor-common/src/main/java/com/win/dfas/monitor/common/vo/MicroServiceJvmMemoryVO.java
@@ -15,6 +15,6 @@ public class MicroServiceJvmMemoryVO {
     private List<String> legendData=new ArrayList<>();
     private List<String> xAxisData=new ArrayList<>();
     private List<String> colorData=new ArrayList<>();
-    private List<List<Integer>> seriesData=new ArrayList<>();
+    private List<List<Double>> seriesData=new ArrayList<>();
 
 }
